[metasrv] refactor: avoid re-export
