+++ metasrv/src/meta_service/raftmeta_test.rs
@@ -25,13 +25,13 @@ use maplit::btreeset;
 use pretty_assertions::assert_eq;
 
 use crate::configs;
-use crate::meta_service::AppliedState;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
 use crate::meta_service::MetaNode;
 use crate::meta_service::NodeId;
 use crate::meta_service::RaftTxId;
 use crate::meta_service::RetryableError;
+use crate::raft::state_machine::AppliedState;
 use crate::tests::assert_meta_connection;
 use crate::tests::service::new_test_context;
 use crate::tests::service::MetaSrvTestContext;
