+++ metasrv/src/raft/state_machine/state_machine_test.rs
@@ -34,9 +34,9 @@ use pretty_assertions::assert_eq;
 use crate::meta_service::testing::pretty_snapshot;
 use crate::meta_service::testing::pretty_snapshot_iter;
 use crate::meta_service::testing::snapshot_logs;
-use crate::meta_service::AppliedState;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
+use crate::raft::state_machine::AppliedState;
 use crate::raft::state_machine::Node;
 use crate::raft::state_machine::Replication;
 use crate::raft::state_machine::SerializableSnapshot;
