+++ store/src/executor/kv_handlers.rs
@@ -24,9 +24,9 @@ use common_store_api_sdk::kv_api_impl::PrefixListReply;
 use common_store_api_sdk::kv_api_impl::PrefixListReq;
 use common_store_api_sdk::kv_api_impl::UpsertKVAction;
 use common_store_api_sdk::kv_api_impl::UpsertKVActionResult;
-use metasrv::meta_service::AppliedState;
 use metasrv::meta_service::Cmd;
 use metasrv::meta_service::LogEntry;
+use metasrv::raft::state_machine::AppliedState;
 
 use crate::executor::action_handler::RequestHandler;
 use crate::executor::ActionHandler;
