+++ store/src/executor/storage_handlers.rs
@@ -19,9 +19,9 @@ use common_store_api_sdk::storage_api_impl::ReadPlanResult;
 use common_store_api_sdk::storage_api_impl::TruncateTableAction;
 use common_store_api_sdk::storage_api_impl::TruncateTableResult;
 use log::debug;
-use metasrv::meta_service::AppliedState;
 use metasrv::meta_service::Cmd;
 use metasrv::meta_service::LogEntry;
+use metasrv::raft::state_machine::AppliedState;
 
 use crate::executor::action_handler::RequestHandler;
 use crate::executor::ActionHandler;
